Specify BUNDLE_GEMFILE in reload task
We specify the BUNDLE_GEMFILE using the symlinked "current_path" so
that unicorn refreshes the gem list properly when receiving the usr2
signal. Otherwise new gems can be missed on a deploy.

An earlier patch adds this fix to the start task. This one completes
the fix by adding it to the reload task.