Change variable to `unicorn_user`.
Don't require `use_sudo` to be true because that has wide-ranging
side-effects on the rest of the Capistrano script. Instead, since sudo
is required for running as a different user, just use `sudo` instead of
`try_sudo`.

Don't fall back on `runner` variable. If someone wants it to be the same
as `runner` they can just set it to be the same.